Ft: revenew update
