+++ src/plugins/wompi-subscription/api/index.ts
@@ -1,3 +1,6 @@
 export * from './wompi-webhook.controller';
+export * from './wompi-token.controller';
 export * from './api-extensions';
-export * from './wompi-subscription.resolver';
+export * from './plan.resolver';
+export * from './subscription.resolver';
+export * from './wompi.resolver';
