+++ src/plugins/wompi-subscription/guards/index.ts
@@ -1,3 +1,5 @@
 export * from './feature.guard';
+export * from './product-limit.guard';
+export * from './product-variation-limit.guard';
+export * from './feature-access.guard';
 export * from './plan.guard';
-
