+++ src/plugins/wompi-subscription/services/index.ts
@@ -1,3 +1,9 @@
 export * from './wompi.service';
-export * from './subscription.service';
+export * from './plan-management.service';
+export * from './subscription-query.service';
+export * from './subscription-write.service';
+export * from './subscription-lifecycle.service';
+export * from './feature-check.service';
+export * from './product-limit-enforcement.service';
 export * from './billing-job.service';
+export * from './billing-email.service';
