--- src/plugins/wompi-subscription/services/subscription.service.ts
@@ -1,833 +1,0 @@
-import { Injectable, OnModuleInit } from '@nestjs/common';
-import { InjectRepository } from '@nestjs/typeorm';
-import { Repository, LessThanOrEqual, MoreThan, IsNull } from 'typeorm';
-import { Logger, Administrator, Product, ProductVariant } from '@vendure/core';
-import { Plan, BillingInterval } from '../entities/plan.entity';
-import { Feature, FeatureType } from '../entities/feature.entity';
-import { PlanFeature } from '../entities/plan-feature.entity';
-import { CustomerSubscription, SubscriptionStatus } from '../entities/customer-subscription.entity';
-import { WompiService } from './wompi.service';
-import {
-    FEATURE_CODES,
-    GRACE_PERIOD_DAYS,
-    DEFAULT_PLAN_NAMES,
-    MANUAL_RENEWAL_REMINDER_DAYS,
-    PaymentFlowType,
-} from '../constants';
-
-@Injectable()
-export class SubscriptionService implements OnModuleInit {
-    constructor(
-        @InjectRepository(Plan) private planRepository: Repository<Plan>,
-        @InjectRepository(Feature) private featureRepository: Repository<Feature>,
-        @InjectRepository(PlanFeature) private planFeatureRepository: Repository<PlanFeature>,
-        @InjectRepository(CustomerSubscription) private subscriptionRepository: Repository<CustomerSubscription>,
-        @InjectRepository(Product) private productRepository: Repository<Product>,
-        @InjectRepository(ProductVariant) private variantRepository: Repository<ProductVariant>,
-        private wompiService: WompiService,
-    ) { }
-
-    async onModuleInit() {
-        await this.ensureDefaultPlansExist();
-    }
-
-    async ensureDefaultPlansExist(): Promise<void> {
-        const existingFree = await this.getPlanByName(DEFAULT_PLAN_NAMES.FREE);
-        if (!existingFree) {
-            await this.createDefaultPlans();
-            Logger.info('Seeded default subscription plans on startup', 'SubscriptionService');
-        }
-    }
-
-    async assignFreePlanToAdministrator(administratorId: number): Promise<CustomerSubscription> {
-        const existing = await this.getSubscriptionByAdministratorId(administratorId);
-        if (existing) {
-            return existing;
-        }
-
-        const freePlan = await this.getFreePlan();
-        const subscription = this.subscriptionRepository.create({
-            administratorId,
-            planId: freePlan.id,
-            status: SubscriptionStatus.ACTIVE,
-            startsAt: new Date(),
-            endsAt: null,
-            autoRenew: false,
-            paymentFlowType: PaymentFlowType.MANUAL,
-        });
-
-        const saved = await this.subscriptionRepository.save(subscription);
-        Logger.info(`Assigned free plan to administrator ${administratorId}`, 'SubscriptionService');
-        return saved;
-    }
-
-    async syncPlanToAllSubscribers(planId: number): Promise<number> {
-        const subscriptions = await this.subscriptionRepository.find({
-            where: { planId },
-        });
-
-        for (const sub of subscriptions) {
-            sub.updatedAt = new Date();
-            await this.subscriptionRepository.save(sub);
-        }
-
-        Logger.info(`Synced ${subscriptions.length} subscriptions to plan ${planId}`, 'SubscriptionService');
-        return subscriptions.length;
-    }
-
-    async getPlanById(planId: number): Promise<Plan | null> {
-        return this.planRepository.findOne({ where: { id: planId }, relations: ['planFeatures', 'planFeatures.feature'] });
-    }
-
-    async getPlanByName(name: string): Promise<Plan | null> {
-        return this.planRepository.findOne({ where: { name }, relations: ['planFeatures', 'planFeatures.feature'] });
-    }
-
-    async getAllPlans(): Promise<Plan[]> {
-        return this.planRepository.find({ relations: ['planFeatures', 'planFeatures.feature'] });
-    }
-
-    async getFreePlan(): Promise<Plan> {
-        let plan = await this.getPlanByName(DEFAULT_PLAN_NAMES.FREE);
-        if (!plan) {
-            plan = await this.createDefaultPlans();
-        }
-        return plan;
-    }
-
-    private async reloadSubscriptionWithPlan(id: number): Promise<CustomerSubscription | null> {
-        return this.subscriptionRepository
-            .createQueryBuilder('sub')
-            .leftJoinAndSelect('sub.plan', 'plan')
-            .leftJoinAndSelect('plan.planFeatures', 'planFeatures')
-            .leftJoinAndSelect('planFeatures.feature', 'feature')
-            .where('sub.id = :id', { id })
-            .getOne();
-    }
-
-    async getSubscriptionByAdministratorId(administratorId: number): Promise<CustomerSubscription | null> {
-        return this.subscriptionRepository
-            .createQueryBuilder('sub')
-            .leftJoinAndSelect('sub.plan', 'plan')
-            .leftJoinAndSelect('plan.planFeatures', 'planFeatures')
-            .leftJoinAndSelect('planFeatures.feature', 'feature')
-            .where('sub.administratorId = :adminId', { adminId: administratorId })
-            .getOne();
-    }
-
-    async getSubscriptionById(subscriptionId: number): Promise<CustomerSubscription | null> {
-        return this.subscriptionRepository
-            .createQueryBuilder('sub')
-            .leftJoinAndSelect('sub.plan', 'plan')
-            .leftJoinAndSelect('plan.planFeatures', 'planFeatures')
-            .leftJoinAndSelect('planFeatures.feature', 'feature')
-            .where('sub.id = :id', { id: subscriptionId })
-            .getOne();
-    }
-
-    async findByPendingReference(reference: string): Promise<CustomerSubscription | null> {
-        return this.subscriptionRepository.findOne({
-            where: { pendingPaymentReference: reference, status: SubscriptionStatus.PENDING_PAYMENT },
-            relations: ['plan'],
-        });
-    }
-
-    async createRecurrentSubscription(
-        administratorId: number,
-        planId: number,
-        paymentMethodType: string,
-        billingPaymentSourceId: string,
-        billingCustomerEmail: string,
-    ): Promise<CustomerSubscription> {
-        const existing = await this.getSubscriptionByAdministratorId(administratorId);
-
-        const plan = await this.getPlanById(planId);
-        if (!plan) {
-            throw new Error('Plan not found');
-        }
-
-        if (existing) {
-            existing.plan = plan;
-            existing.planId = planId;
-            existing.paymentMethodType = paymentMethodType;
-            existing.billingPaymentSourceId = billingPaymentSourceId;
-            existing.billingCustomerEmail = billingCustomerEmail;
-            existing.status = SubscriptionStatus.ACTIVE;
-            existing.autoRenew = true;
-            existing.paymentFlowType = PaymentFlowType.RECURRENTE;
-            existing.endsAt = this.calculateEndDate(plan.billingInterval, existing.endsAt ?? undefined);
-            existing.lastPaymentAt = new Date();
-            existing.gracePeriodStart = null as any;
-
-            const saved = await this.subscriptionRepository.save(existing);
-            const reloaded = await this.reloadSubscriptionWithPlan(saved.id);
-
-            const productLimitValue = await this.getFeatureValue(administratorId, FEATURE_CODES.MAX_PRODUCTS);
-            if (productLimitValue) {
-                await this.restoreHiddenProducts(administratorId, parseInt(productLimitValue, 10));
-            }
-
-            const variantLimitValue = await this.getFeatureValue(administratorId, FEATURE_CODES.MAX_VARIATIONS);
-            if (variantLimitValue) {
-                await this.restoreHiddenVariants(administratorId, parseInt(variantLimitValue, 10));
-            }
-
-            Logger.info(`Upgraded subscription ${saved.id} for administrator ${administratorId} to plan ${plan.name}`, 'SubscriptionService');
-            return reloaded ?? saved;
-        }
-
-        const subscription = this.subscriptionRepository.create({
-            administratorId,
-            planId,
-            status: SubscriptionStatus.ACTIVE,
-            startsAt: new Date(),
-            endsAt: this.calculateEndDate(plan.billingInterval),
-            autoRenew: true,
-            billingPaymentSourceId,
-            billingCustomerEmail,
-            paymentMethodType,
-            paymentFlowType: PaymentFlowType.RECURRENTE,
-            lastPaymentAt: new Date(),
-        });
-
-        const saved = await this.subscriptionRepository.save(subscription);
-        const reloaded = await this.reloadSubscriptionWithPlan(saved.id);
-        Logger.info(`Created recurrent subscription ${saved.id} for administrator ${administratorId} with plan ${plan.name}`, 'SubscriptionService');
-        return reloaded ?? saved;
-    }
-
-    async createPendingSubscription(
-        administratorId: number,
-        planId: number,
-        paymentMethodType: string,
-    ): Promise<{ subscription: CustomerSubscription; reference: string }> {
-        const existing = await this.getSubscriptionByAdministratorId(administratorId);
-
-        const plan = await this.getPlanById(planId);
-        if (!plan) {
-            throw new Error('Plan not found');
-        }
-
-        const reference = `SUB-PENDING-${administratorId}-${Date.now()}`;
-
-        if (existing) {
-            existing.plan = plan;
-            existing.planId = planId;
-            existing.paymentMethodType = paymentMethodType;
-            existing.status = SubscriptionStatus.PENDING_PAYMENT;
-            existing.paymentFlowType = PaymentFlowType.MANUAL;
-            existing.pendingPaymentReference = reference;
-
-            const saved = await this.subscriptionRepository.save(existing);
-            const reloaded = await this.reloadSubscriptionWithPlan(saved.id);
-            Logger.info(`Upgraded pending subscription ${saved.id} for administrator ${administratorId}`, 'SubscriptionService');
-            return { subscription: reloaded ?? saved, reference };
-        }
-
-        const subscription = this.subscriptionRepository.create({
-            administratorId,
-            planId,
-            status: SubscriptionStatus.PENDING_PAYMENT,
-            startsAt: new Date(),
-            endsAt: this.calculateEndDate(plan.billingInterval),
-            autoRenew: false,
-            paymentMethodType,
-            paymentFlowType: PaymentFlowType.MANUAL,
-            pendingPaymentReference: reference,
-        });
-
-        const saved = await this.subscriptionRepository.save(subscription);
-        const reloaded = await this.reloadSubscriptionWithPlan(saved.id);
-        Logger.info(`Created pending subscription ${saved.id} for administrator ${administratorId}`, 'SubscriptionService');
-        return { subscription: reloaded ?? saved, reference };
-    }
-
-    async activateSubscriptionAfterPayment(subscriptionId: number, paymentSourceId?: string): Promise<CustomerSubscription> {
-        const subscription = await this.subscriptionRepository.findOne({ where: { id: subscriptionId }, relations: ['plan'] });
-        if (!subscription) {
-            throw new Error('Subscription not found');
-        }
-
-        subscription.status = SubscriptionStatus.ACTIVE;
-        subscription.pendingPaymentReference = null;
-        subscription.lastPaymentAt = new Date();
-        if (paymentSourceId) {
-            subscription.billingPaymentSourceId = paymentSourceId;
-        }
-
-        return this.subscriptionRepository.save(subscription);
-    }
-
-    async createSubscription(
-        administratorId: number,
-        planId: number,
-        billingPaymentSourceId: string,
-        billingCustomerEmail: string,
-    ): Promise<CustomerSubscription> {
-        return this.createRecurrentSubscription(
-            administratorId,
-            planId,
-            'CARD',
-            billingPaymentSourceId,
-            billingCustomerEmail,
-        );
-    }
-
-    async updateSubscriptionStatus(subscriptionId: number, status: SubscriptionStatus): Promise<CustomerSubscription> {
-        const subscription = await this.subscriptionRepository.findOne({ where: { id: subscriptionId }, relations: ['plan'] });
-        if (!subscription) {
-            throw new Error('Subscription not found');
-        }
-
-        subscription.status = status;
-        if (status === SubscriptionStatus.GRACE_PERIOD) {
-            subscription.gracePeriodStart = new Date();
-        } else if (status === SubscriptionStatus.ACTIVE) {
-            subscription.gracePeriodStart = null;
-            subscription.startsAt = new Date();
-            subscription.endsAt = this.calculateEndDate(subscription.plan?.billingInterval || BillingInterval.MONTHLY);
-        }
-
-        return this.subscriptionRepository.save(subscription);
-    }
-
-    async extendSubscription(subscriptionId: number): Promise<CustomerSubscription> {
-        const subscription = await this.subscriptionRepository.findOne({
-            where: { id: subscriptionId },
-            relations: ['plan'],
-        });
-        if (!subscription) {
-            throw new Error('Subscription not found');
-        }
-
-        subscription.endsAt = this.calculateEndDate(subscription.plan.billingInterval, subscription.endsAt ?? undefined);
-        subscription.lastPaymentAt = new Date();
-        return this.subscriptionRepository.save(subscription);
-    }
-
-    async extendManualSubscription(subscriptionId: number): Promise<CustomerSubscription> {
-        const subscription = await this.subscriptionRepository.findOne({
-            where: { id: subscriptionId },
-            relations: ['plan'],
-        });
-        if (!subscription) {
-            throw new Error('Subscription not found');
-        }
-
-        subscription.endsAt = this.calculateEndDate(subscription.plan.billingInterval, subscription.endsAt || new Date());
-        subscription.lastPaymentAt = new Date();
-        return this.subscriptionRepository.save(subscription);
-    }
-
-    async stopAutoRenew(subscriptionId: number): Promise<CustomerSubscription> {
-        const subscription = await this.subscriptionRepository.findOne({ where: { id: subscriptionId } });
-        if (!subscription) {
-            throw new Error('Subscription not found');
-        }
-
-        subscription.autoRenew = false;
-        Logger.info(`Stopped auto-renew for subscription ${subscriptionId}`, 'SubscriptionService');
-        return this.subscriptionRepository.save(subscription);
-    }
-
-    async cancelSubscription(subscriptionId: number): Promise<CustomerSubscription> {
-        const subscription = await this.subscriptionRepository.findOne({
-            where: { id: subscriptionId },
-            relations: ['plan'],
-        });
-        if (!subscription) {
-            throw new Error('Subscription not found');
-        }
-
-        const administratorId = subscription.administratorId;
-
-        if (subscription.billingPaymentSourceId) {
-            await this.wompiService.deletePaymentSource(subscription.billingPaymentSourceId);
-        }
-
-        const freePlan = await this.getFreePlan();
-        subscription.plan = freePlan;
-        subscription.planId = freePlan.id;
-        subscription.status = SubscriptionStatus.ACTIVE;
-        subscription.autoRenew = false;
-        subscription.billingPaymentSourceId = null;
-        subscription.billingCustomerEmail = null as any;
-        subscription.paymentMethodType = null as any;
-        subscription.paymentFlowType = null as any;
-        subscription.gracePeriodStart = null as any;
-        subscription.lastPaymentAt = null as any;
-
-        const saved = await this.subscriptionRepository.save(subscription);
-        const reloaded = await this.reloadSubscriptionWithPlan(saved.id);
-
-        const productLimitValue = await this.getFeatureValue(administratorId, FEATURE_CODES.MAX_PRODUCTS);
-        const productLimit = productLimitValue ? parseInt(productLimitValue, 10) : 15;
-        await this.hideExcessProducts(administratorId, productLimit);
-
-        const variantLimitValue = await this.getFeatureValue(administratorId, FEATURE_CODES.MAX_VARIATIONS);
-        const variantLimit = variantLimitValue ? parseInt(variantLimitValue, 10) : 0;
-        await this.hideExcessVariants(administratorId, variantLimit);
-
-        Logger.info(`Subscription ${subscriptionId} reverted to Free plan for administrator ${administratorId}`, 'SubscriptionService');
-        return reloaded ?? saved;
-    }
-
-    async cancelAutoRenew(administratorId: number): Promise<CustomerSubscription> {
-        const subscription = await this.getSubscriptionByAdministratorId(administratorId);
-        if (!subscription) {
-            throw new Error('No active subscription found');
-        }
-
-        subscription.autoRenew = false;
-        return this.subscriptionRepository.save(subscription);
-    }
-
-    async getFeatureValue(administratorId: number, featureCode: string): Promise<string | null> {
-        const subscription = await this.getSubscriptionByAdministratorId(administratorId);
-        if (!subscription || subscription.status !== SubscriptionStatus.ACTIVE) {
-            return null;
-        }
-
-        const feature = await this.featureRepository.findOne({ where: { code: featureCode } });
-        if (!feature) {
-            return null;
-        }
-
-        const planFeature = await this.planFeatureRepository.findOne({
-            where: { planId: subscription.planId, featureId: feature.id },
-        });
-
-        return planFeature?.value || null;
-    }
-
-    async checkFeatureAccess(administratorId: number, featureCode: string): Promise<boolean> {
-        const value = await this.getFeatureValue(administratorId, featureCode);
-        const feature = await this.featureRepository.findOne({ where: { code: featureCode } });
-
-        if (!feature || !value) {
-            return feature?.type === FeatureType.BOOLEAN ? false : false;
-        }
-
-        if (feature.type === FeatureType.BOOLEAN) {
-            return value.toLowerCase() === 'true';
-        }
-
-        return true;
-    }
-
-    async checkProductLimit(administratorId: number, channelToken?: string): Promise<{ allowed: boolean; current: number; limit: number }> {
-        const subscription = await this.getSubscriptionByAdministratorId(administratorId);
-        if (!subscription || subscription.status !== SubscriptionStatus.ACTIVE) {
-            return { allowed: false, current: 0, limit: 0 };
-        }
-
-        const limitValue = await this.getFeatureValue(administratorId, FEATURE_CODES.MAX_PRODUCTS);
-        const limit = limitValue ? parseInt(limitValue, 10) : 0;
-
-        const channel = channelToken
-            ? await this.findChannelByCode(channelToken) ?? await this.getChannelByAdministratorId(administratorId)
-            : await this.getChannelByAdministratorId(administratorId);
-        if (!channel) {
-            return { allowed: limit > 0, current: 0, limit };
-        }
-
-        const productCount = await this.productRepository
-            .createQueryBuilder('product')
-            .innerJoin('product.channels', 'channel')
-            .where('channel.id = :channelId', { channelId: channel.id })
-            .andWhere('product.deletedAt IS NULL')
-            .andWhere('product."customFieldsHidden" IS DISTINCT FROM true')
-            .getCount();
-
-        return {
-            allowed: productCount < limit,
-            current: productCount,
-            limit,
-        };
-    }
-
-    async checkVariationLimit(administratorId: number, channelToken?: string): Promise<{ allowed: boolean; current: number; limit: number }> {
-        const subscription = await this.getSubscriptionByAdministratorId(administratorId);
-        if (!subscription || subscription.status !== SubscriptionStatus.ACTIVE) {
-            return { allowed: false, current: 0, limit: 0 };
-        }
-
-        const limitValue = await this.getFeatureValue(administratorId, FEATURE_CODES.MAX_VARIATIONS);
-        const limit = limitValue ? parseInt(limitValue, 10) : 0;
-
-        const channel = channelToken
-            ? await this.findChannelByCode(channelToken) ?? await this.getChannelByAdministratorId(administratorId)
-            : await this.getChannelByAdministratorId(administratorId);
-        if (!channel) {
-            return { allowed: limit > 0, current: 0, limit };
-        }
-
-        const variantCount = await this.variantRepository
-            .createQueryBuilder('variant')
-            .innerJoin('variant.product', 'product')
-            .innerJoin('product.channels', 'channel')
-            .where('channel.id = :channelId', { channelId: channel.id })
-            .andWhere('product.deletedAt IS NULL')
-            .andWhere('variant.deletedAt IS NULL')
-            .andWhere('variant.enabled = :enabled', { enabled: true })
-            .andWhere('variant."customFieldsHidden" IS DISTINCT FROM true')
-            .getCount();
-
-        return {
-            allowed: variantCount < limit,
-            current: variantCount,
-            limit,
-        };
-    }
-
-    private async findChannelByCode(code: string): Promise<any> {
-        const { Channel } = await import('@vendure/core');
-        const repo = this.planRepository.manager.getRepository(Channel);
-        return repo.findOne({ where: { code } });
-    }
-
-    private async getChannelByAdministratorId(administratorId: number): Promise<any> {
-        const { Channel } = await import('@vendure/core');
-        const admin = await this.planRepository.manager.findOne(Administrator, {
-            where: { id: administratorId },
-            relations: ['user', 'user.roles', 'user.roles.channels'],
-        });
-        if (!admin) return null;
-        for (const role of admin.user.roles) {
-            const sellerChannel = role.channels.find((ch: any) => ch.sellerId != null);
-            if (sellerChannel) return sellerChannel;
-        }
-        return null;
-    }
-
-    async getActiveSubscriptionsForRenewal(): Promise<CustomerSubscription[]> {
-        const today = new Date();
-        today.setHours(0, 0, 0, 0);
-
-        return this.subscriptionRepository.find({
-            where: {
-                status: SubscriptionStatus.ACTIVE,
-                autoRenew: true,
-                endsAt: LessThanOrEqual(today),
-            },
-            relations: ['plan'],
-        });
-    }
-
-    async getGracePeriodSubscriptions(): Promise<CustomerSubscription[]> {
-        const cutoffDate = new Date();
-        cutoffDate.setDate(cutoffDate.getDate() - GRACE_PERIOD_DAYS);
-
-        return this.subscriptionRepository.find({
-            where: {
-                status: SubscriptionStatus.GRACE_PERIOD,
-                gracePeriodStart: LessThanOrEqual(cutoffDate),
-            },
-            relations: ['plan'],
-        });
-    }
-
-    async getSuspendedSubscriptionsForPurge(): Promise<CustomerSubscription[]> {
-        const cutoffDate = new Date();
-        cutoffDate.setDate(cutoffDate.getDate() - (GRACE_PERIOD_DAYS * 2));
-
-        return this.subscriptionRepository.find({
-            where: {
-                status: SubscriptionStatus.SUSPENDED,
-                gracePeriodStart: LessThanOrEqual(cutoffDate),
-            },
-            relations: ['plan'],
-        });
-    }
-
-    async getManualSubscriptionsDueForRenewal(): Promise<CustomerSubscription[]> {
-        const futureDate = new Date();
-        futureDate.setDate(futureDate.getDate() + MANUAL_RENEWAL_REMINDER_DAYS);
-
-        const today = new Date();
-        today.setHours(0, 0, 0, 0);
-
-        return this.subscriptionRepository.find({
-            where: {
-                status: SubscriptionStatus.ACTIVE,
-                autoRenew: false,
-                paymentFlowType: PaymentFlowType.MANUAL,
-                endsAt: LessThanOrEqual(futureDate),
-            },
-            relations: ['plan'],
-        });
-    }
-
-    async getPendingPaymentSubscriptions(): Promise<CustomerSubscription[]> {
-        const cutoffDate = new Date();
-        cutoffDate.setHours(cutoffDate.getHours() - 24);
-
-        return this.subscriptionRepository.find({
-            where: {
-                status: SubscriptionStatus.PENDING_PAYMENT,
-                createdAt: LessThanOrEqual(cutoffDate),
-            },
-            relations: ['plan'],
-        });
-    }
-
-    async downgradeToFree(subscriptionId: number): Promise<CustomerSubscription> {
-        const freePlan = await this.getFreePlan();
-        const subscription = await this.subscriptionRepository.findOne({ where: { id: subscriptionId } });
-
-        if (!subscription) {
-            throw new Error('Subscription not found');
-        }
-
-        subscription.plan = freePlan;
-        subscription.planId = freePlan.id;
-        subscription.status = SubscriptionStatus.ACTIVE;
-
-        return this.subscriptionRepository.save(subscription);
-    }
-
-    async hideExcessProducts(administratorId: number, maxAllowed: number): Promise<number> {
-        const channel = await this.getChannelByAdministratorId(administratorId);
-        if (!channel) {
-            Logger.warn(`No channel found for administrator ${administratorId}`, 'SubscriptionService');
-            return 0;
-        }
-
-        const products = await this.productRepository.find({
-            where: {
-                channels: { id: channel.id },
-                deletedAt: IsNull(),
-            },
-            relations: ['channels'],
-            order: { createdAt: 'ASC' },
-            take: 1000,
-        });
-
-        const activeProducts = products.filter(p => !(p as any).customFields?.hidden);
-        const productsToHide = activeProducts.slice(maxAllowed);
-        let hiddenCount = 0;
-
-        for (const product of productsToHide) {
-            const existingCustomFields = (product as any).customFields || {};
-            (product as any).customFields = {
-                ...existingCustomFields,
-                hidden: true,
-                hiddenAt: new Date(),
-            };
-            await this.productRepository.save(product);
-            hiddenCount++;
-        }
-
-        Logger.info(`Hidden ${hiddenCount} products for administrator ${administratorId}`, 'SubscriptionService');
-        return hiddenCount;
-    }
-
-    async restoreHiddenProducts(administratorId: number, maxAllowed: number): Promise<number> {
-        const channel = await this.getChannelByAdministratorId(administratorId);
-        if (!channel) {
-            Logger.warn(`No channel found for administrator ${administratorId}`, 'SubscriptionService');
-            return 0;
-        }
-
-        const allProducts = await this.productRepository.find({
-            where: {
-                channels: { id: channel.id },
-            },
-            relations: ['channels'],
-            order: { createdAt: 'ASC' },
-            take: 1000,
-        });
-
-        const activeCount = allProducts.filter(
-            p => !(p as any).customFields?.hidden && !p.deletedAt
-        ).length;
-
-        const hiddenProducts = allProducts.filter(
-            p => (p as any).customFields?.hidden && !p.deletedAt
-        );
-
-        const canRestore = Math.max(0, maxAllowed - activeCount);
-        if (canRestore <= 0) return 0;
-
-        const toRestore = hiddenProducts.slice(0, canRestore);
-        let restoredCount = 0;
-
-        for (const product of toRestore) {
-            const existingCustomFields = (product as any).customFields || {};
-            (product as any).customFields = {
-                ...existingCustomFields,
-                hidden: false,
-                hiddenAt: null,
-            };
-            await this.productRepository.save(product);
-            restoredCount++;
-        }
-
-        Logger.info(`Restored ${restoredCount} products for administrator ${administratorId}`, 'SubscriptionService');
-        return restoredCount;
-    }
-
-    async hideExcessVariants(administratorId: number, maxAllowed: number): Promise<number> {
-        const channel = await this.getChannelByAdministratorId(administratorId);
-        if (!channel) {
-            Logger.warn(`No channel found for administrator ${administratorId}`, 'SubscriptionService');
-            return 0;
-        }
-
-        const variants = await this.variantRepository.find({
-            where: {
-                product: {
-                    channels: { id: channel.id },
-                    deletedAt: IsNull(),
-                },
-                enabled: true,
-                deletedAt: IsNull(),
-            },
-            relations: ['product', 'product.channels'],
-            order: { createdAt: 'ASC' },
-            take: 10000,
-        });
-
-        const activeVariants = variants.filter(v => !(v as any).customFields?.hidden);
-        const variantsToHide = activeVariants.slice(maxAllowed);
-        let hiddenCount = 0;
-
-        for (const variant of variantsToHide) {
-            const existingCustomFields = (variant as any).customFields || {};
-            (variant as any).customFields = {
-                ...existingCustomFields,
-                hidden: true,
-                hiddenAt: new Date(),
-            };
-            await this.variantRepository.save(variant);
-            hiddenCount++;
-        }
-
-        Logger.info(`Hidden ${hiddenCount} variants for administrator ${administratorId}`, 'SubscriptionService');
-        return hiddenCount;
-    }
-
-    async restoreHiddenVariants(administratorId: number, maxAllowed: number): Promise<number> {
-        const channel = await this.getChannelByAdministratorId(administratorId);
-        if (!channel) {
-            Logger.warn(`No channel found for administrator ${administratorId}`, 'SubscriptionService');
-            return 0;
-        }
-
-        const allVariants = await this.variantRepository.find({
-            where: {
-                product: {
-                    channels: { id: channel.id },
-                },
-                deletedAt: IsNull(),
-            },
-            relations: ['product', 'product.channels'],
-            order: { createdAt: 'ASC' },
-            take: 10000,
-        });
-
-        const activeCount = allVariants.filter(
-            v => !(v as any).customFields?.hidden && v.enabled
-        ).length;
-
-        const hiddenVariants = allVariants.filter(
-            v => (v as any).customFields?.hidden
-        );
-
-        const canRestore = Math.max(0, maxAllowed - activeCount);
-        if (canRestore <= 0) return 0;
-
-        const toRestore = hiddenVariants.slice(0, canRestore);
-        let restoredCount = 0;
-
-        for (const variant of toRestore) {
-            const existingCustomFields = (variant as any).customFields || {};
-            (variant as any).customFields = {
-                ...existingCustomFields,
-                hidden: false,
-                hiddenAt: null,
-            };
-            await this.variantRepository.save(variant);
-            restoredCount++;
-        }
-
-        Logger.info(`Restored ${restoredCount} variants for administrator ${administratorId}`, 'SubscriptionService');
-        return restoredCount;
-    }
-
-    private calculateEndDate(interval: BillingInterval, fromDate?: Date): Date {
-        const startDate = fromDate || new Date();
-        const endDate = new Date(startDate);
-
-        if (interval === BillingInterval.MONTHLY) {
-            endDate.setMonth(endDate.getMonth() + 1);
-        } else if (interval === BillingInterval.YEARLY) {
-            endDate.setFullYear(endDate.getFullYear() + 1);
-        }
-
-        return endDate;
-    }
-
-    async createDefaultPlans(): Promise<Plan> {
-        const freePlan = this.planRepository.create({
-            name: DEFAULT_PLAN_NAMES.FREE,
-            price: 0,
-            billingInterval: BillingInterval.MONTHLY,
-            isActive: true,
-            description: 'Plan gratuito con características limitadas',
-        });
-        const savedFreePlan = await this.planRepository.save(freePlan);
-
-        const tiendaPlan = this.planRepository.create({
-            name: DEFAULT_PLAN_NAMES.TIENDA,
-            price: 29900,
-            billingInterval: BillingInterval.MONTHLY,
-            isActive: true,
-            description: 'Plan para tiendas con hasta 500 productos',
-        });
-        const savedTiendaPlan = await this.planRepository.save(tiendaPlan);
-
-        const omnichannelPlan = this.planRepository.create({
-            name: DEFAULT_PLAN_NAMES.OMNICHANNEL,
-            price: 99900,
-            billingInterval: BillingInterval.MONTHLY,
-            isActive: true,
-            description: 'Plan multicanal con hasta 1.500 productos',
-        });
-        const savedOmnichannelPlan = await this.planRepository.save(omnichannelPlan);
-
-        const features = [
-            { code: FEATURE_CODES.MAX_PRODUCTS, name: 'Max Products', type: FeatureType.NUMERIC },
-            { code: FEATURE_CODES.MAX_VARIATIONS, name: 'Max Variations', type: FeatureType.NUMERIC },
-            { code: FEATURE_CODES.AI_ACCESS, name: 'AI Access', type: FeatureType.BOOLEAN },
-            { code: FEATURE_CODES.ELECTRONIC_BILLING, name: 'Electronic Billing', type: FeatureType.BOOLEAN },
-        ];
-
-        const planConfigs = [
-            { planId: savedFreePlan.id, values: { max_products: '15', max_variations: '250', ai_access: 'false', electronic_billing: 'false' } },
-            { planId: savedTiendaPlan.id, values: { max_products: '500', max_variations: '5000', ai_access: 'true', electronic_billing: 'true' } },
-            { planId: savedOmnichannelPlan.id, values: { max_products: '1500', max_variations: '15000', ai_access: 'true', electronic_billing: 'true' } },
-        ];
-
-        for (const featureData of features) {
-            let feature = await this.featureRepository.findOne({ where: { code: featureData.code } });
-            if (!feature) {
-                feature = this.featureRepository.create(featureData);
-                feature = await this.featureRepository.save(feature);
-            }
-
-            for (const config of planConfigs) {
-                await this.planFeatureRepository.save(
-                    this.planFeatureRepository.create({
-                        planId: config.planId,
-                        featureId: feature.id,
-                        value: config.values[featureData.code as keyof typeof config.values],
-                    })
-                );
-            }
-        }
-
-        Logger.info('Created default subscription plans', 'SubscriptionService');
-        return savedFreePlan;
-    }
-}
