+++ static/email/templates/subscription/manual-reminder.hbs
@@ -1,0 +1,25 @@
+{{> header title="Recordatorio de pago" }}
+
+<mj-section padding="20px">
+  <mj-column background-color="#ffffff" padding="30px" border-radius="12px">
+
+    <mj-text font-size="20px" font-weight="bold" color="#12123F">
+      Recordatorio de pago
+    </mj-text>
+
+    <mj-text>
+      Tu plan <strong>{{planName}}</strong> vence en <strong>{{daysLeft}} días</strong>.
+    </mj-text>
+
+    <mj-text>
+      Realiza el pago a tiempo para evitar la suspensión de tu plan y la desactivación de productos.
+    </mj-text>
+
+    <mj-text>
+      Ingresa al panel de administración para gestionar tu suscripción.
+    </mj-text>
+
+  </mj-column>
+</mj-section>
+
+{{> footer }}
