+++ static/email/templates/subscription/payment-expired.hbs
@@ -1,0 +1,21 @@
+{{> header title="Pago pendiente expirado" }}
+
+<mj-section padding="20px">
+  <mj-column background-color="#ffffff" padding="30px" border-radius="12px">
+
+    <mj-text font-size="20px" font-weight="bold" color="#12123F">
+      Pago pendiente expirado
+    </mj-text>
+
+    <mj-text>
+      El pago pendiente de tu plan <strong>{{planName}}</strong> ha expirado.
+    </mj-text>
+
+    <mj-text>
+      Tu cuenta ha sido movida a período de gracia. Tienes 15 días para completar el pago antes de que tu plan sea suspendido.
+    </mj-text>
+
+  </mj-column>
+</mj-section>
+
+{{> footer }}
