+++ static/email/templates/subscription/renewal-success.hbs
@@ -1,0 +1,21 @@
+{{> header title="Renovación exitosa" }}
+
+<mj-section padding="20px">
+  <mj-column background-color="#ffffff" padding="30px" border-radius="12px">
+
+    <mj-text font-size="20px" font-weight="bold" color="#12123F">
+      Renovación exitosa
+    </mj-text>
+
+    <mj-text>
+      Tu plan <strong>{{planName}}</strong> ha sido renovado automáticamente.
+    </mj-text>
+
+    <mj-text>
+      <strong>Próximo vencimiento:</strong> {{endsAt}}
+    </mj-text>
+
+  </mj-column>
+</mj-section>
+
+{{> footer }}
