+++ static/email/templates/subscription/suspended.hbs
@@ -1,0 +1,21 @@
+{{> header title="Plan suspendido" }}
+
+<mj-section padding="20px">
+  <mj-column background-color="#ffffff" padding="30px" border-radius="12px">
+
+    <mj-text font-size="20px" font-weight="bold" color="#12123F">
+      Plan suspendido
+    </mj-text>
+
+    <mj-text>
+      Tu plan <strong>{{planName}}</strong> ha sido suspendido por falta de pago.
+    </mj-text>
+
+    <mj-text>
+      Los productos que exceden el límite del plan gratuito han sido desactivados. Si renuevas tu plan, los productos se restaurarán automáticamente.
+    </mj-text>
+
+  </mj-column>
+</mj-section>
+
+{{> footer }}
